#39 - refactor: 날짜변환함수 분리
